client_reserv/serializers.py
from rest_framework import serializers
from client_reserv.models import Reservation, TimeSlot


class ReservationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Reservation
        fields = '__all__'


class TimeSlotSerializer(serializers.ModelSerializer):
    class Meta:
        model = TimeSlot
        fields = '__all__'
