requirements.txt
Django
django-cors-headers
djangorestframework
pytest
whitenoise